[GTKUI] Fix date columns to use int not float

+++ deluge/ui/gtkui/torrentview.py
@@ -267,11 +267,11 @@ class TorrentView(ListView, component.Component):
                              status_field=["ratio"], default=False)
         self.add_func_column(_("Avail"), funcs.cell_data_ratio_avail, [float],
                              status_field=["distributed_copies"], default=False)
-        self.add_func_column(_("Added"), funcs.cell_data_date_added, [float],
+        self.add_func_column(_("Added"), funcs.cell_data_date_added, [int],
                              status_field=["time_added"], default=False)
-        self.add_func_column(_("Completed"), funcs.cell_data_date_completed, [float],
+        self.add_func_column(_("Completed"), funcs.cell_data_date_completed, [int],
                              status_field=["completed_time"], default=False)
-        self.add_func_column(_("Complete Seen"), funcs.cell_data_date_or_never, [float],
+        self.add_func_column(_("Complete Seen"), funcs.cell_data_date_or_never, [int],
                              status_field=["last_seen_complete"], default=False)
         self.add_texticon_column(_("Tracker"), function=funcs.cell_data_trackericon,
                                  status_field=["tracker_host", "tracker_host"], default=False)
